Separate examples from word handling
